# specs/002-qabul-application/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Qabul Application Management API
  version: 0.1.0
  description: REST contract for application submission, moderation, and auto-accept control.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/applications:
    post:
      summary: Submit a new application
      operationId: submitApplication
      tags: [Applications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      summary: List applications with pagination and filters
      operationId: listApplications
      tags: [Applications]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ApplicationStatus'
        - name: school_year
          in: query
          schema:
            type: integer
        - name: major_id
          in: query
          schema:
            type: integer
        - name: major_type_id
          in: query
          schema:
            type: integer
        - name: major_lang_id
          in: query
          schema:
            type: integer
        - name: degree
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated application list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/applications/me:
    get:
      summary: Get current applicant application
      operationId: getMyApplication
      tags: [Applications]
      responses:
        '200':
          description: Applicant application found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/applications/{applicationId}:
    get:
      summary: Get application details by id
      operationId: getApplicationById
      tags: [Applications]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/applications/{applicationId}/accept:
    post:
      summary: Accept application
      operationId: acceptApplication
      tags: [Applications]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Application accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/applications/{applicationId}/reject:
    post:
      summary: Reject (cancel) application
      operationId: rejectApplication
      tags: [Applications]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Application rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/applications/settings/auto-accept:
    patch:
      summary: Toggle auto-accept for new submissions
      operationId: toggleAutoAccept
      tags: [Application Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoAcceptToggleRequest'
      responses:
        '200':
          description: Toggle applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutoAcceptToggleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request data or invalid state transition
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: User lacks required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with current state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ApplicationStatus:
      type: string
      enum: [PENDING, IN_REVIEW, ACCEPTED, CANCELED]
    Degree:
      type: string
      enum: [BACHELOR, MASTER]
    CreateApplicationRequest:
      type: object
      required:
        - firstname
        - lastname
        - birth_date
        - gender
        - jshshir
        - passport_series
        - address
        - tuition_id
        - certificates
      properties:
        firstname:
          type: string
        lastname:
          type: string
        middlename:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [M, F]
        jshshir:
          type: string
        passport_series:
          type: string
        address:
          type: string
        additional_phone:
          type: string
        disability:
          type: integer
          default: 0
        tuition_id:
          type: string
          format: uuid
        certificates:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CertificateInput'
    CertificateInput:
      type: object
      required: [cert_number, score, file_url, category_id]
      properties:
        cert_number:
          type: string
        score:
          type: number
          format: double
        file_url:
          type: string
          format: uri
        category_id:
          type: integer
    ApplicationResponse:
      type: object
      required:
        - id
        - user_id
        - tuition_id
        - status
        - firstname
        - lastname
        - birth_date
        - gender
        - jshshir
        - passport_series
        - address
        - disability
        - certificates
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        tuition_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        firstname:
          type: string
        lastname:
          type: string
        middlename:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
          enum: [M, F]
        jshshir:
          type: string
        passport_series:
          type: string
        address:
          type: string
        additional_phone:
          type: string
        disability:
          type: integer
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/CertificateResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CertificateResponse:
      type: object
      required: [id, cert_number, score, file_url, category_id]
      properties:
        id:
          type: string
          format: uuid
        cert_number:
          type: string
        score:
          type: number
          format: double
        file_url:
          type: string
        category_id:
          type: integer
    ApplicationPageResponse:
      type: object
      required: [page, limit, total, data]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResponse'
    AutoAcceptToggleRequest:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    AutoAcceptToggleResponse:
      type: object
      required: [enabled, updated_at]
      properties:
        enabled:
          type: boolean
        updated_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time